backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.20
openai==1.57.2
pdfplumber==0.11.4
python-dotenv==1.0.1
